adding caseUtil
